Update nav file structure
