--- src/components/layout/nav/index.ts
@@ -1,3 +1,0 @@
-export * from './nav';
-export * from './nav-link';
-export * from './nav-links';
