+++ src/components/layout/nav/links.tsx
@@ -1,8 +1,12 @@
 import type { ComponentPropsWithoutRef } from 'react';
 import Link from 'next/link';
+import { MDXContent } from '@content-collections/mdx/react';
 
 import { Button } from '@/components/ui/button';
 import { cn } from '@/lib/cn';
+import { allDocs, navLinksName } from '@/constants/docs';
+
+import { mdxComponents } from '../mdx';
 
 export const NavLink = ({
   href,
@@ -24,4 +28,16 @@ export const NavLink = ({
       </LinkComp>
     </Button>
   );
+};
+
+export const NavLinks = () => {
+  // TODO: Remove this, move links to config.json
+  const linksDoc = allDocs.find(doc => doc._meta.path === navLinksName);
+  if (linksDoc)
+    return (
+      <MDXContent
+        code={linksDoc.body.code}
+        components={{ ...mdxComponents, NavLink }}
+      />
+    );
 };
